html/semantics/embedded-content/media-elements/loading-the-media-resource/resource-selection-pointer-move-non-source-pointer-after-out.html

<!doctype html>
<title>pointer updates (moveBefore non-source pointerAfter out of the media element)</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<div id="log"></div>
<div id="holder"></div>
<script>
var events = [];
var a = 0;
var b = 0;
var c = 0;
</script>
<video id="v"><source id="source-a" onerror="a++; events.push('a')"><source id="source-b" onerror="b++; events.push('b')" src="resources/delayed-broken-video.py">x<source id="source-c" onerror="c++; events.push('c')"></video>
<script>
async_test(function(t) {
  var video = document.getElementById('v');
  assert_implements(typeof video.moveBefore == 'function', 'ParentNode.moveBefore() is required for this test');

  var text = null;
  for (var i = 0; i < video.childNodes.length; ++i) {
    if (video.childNodes[i].nodeType == Node.TEXT_NODE && video.childNodes[i].data == 'x') {
      text = video.childNodes[i];
      break;
    }
  }
  assert_not_equals(text, null, 'text node pointerAfter exists');

  // Move the text node that is pointerAfter out while b is pending.
  document.getElementById('holder').moveBefore(text, null);

  window.onload = t.step_func(function() {
    assert_equals(a, 1, 'error events on a');
    assert_equals(b, 1, 'error events on b');
    assert_equals(c, 1, 'error events on c');
    assert_array_equals(events, ['a', 'b', 'c'], 'processed source order');
    t.done();
  });
});
</script>
